// notaryDapp-Frontend/src/app/models/custumer.model.ts
export class Custumer {
    wallet?: string;
    custumerType?: string;
    address?:{
      address?:string,
      city?: string,
      country?: string
        };
    phone?: string;
    fullNAme?: string;
    gender?: string;
    identityDoc?: string;
    identityCode?: string;
    nationality?: string;
    birthday?: string;
    placeOfBirth?: string
  }